Implemented Data Store Component.
